refactor: move to conversion.py
