+++ python/pyspark/sql/tests/test_conversion.py
@@ -23,6 +23,7 @@ from pyspark.sql.conversion import (
     ArrowTableToRowsConversion,
     LocalDataToArrowConversion,
     ArrowTimestampConversion,
+    ArrowBatchTransformer,
 )
 from pyspark.sql.types import (
     ArrayType,
@@ -62,6 +63,85 @@ class Score:
 
     def __eq__(self, other):
         return self.score == other.score
+
+
+@unittest.skipIf(not have_pyarrow, pyarrow_requirement_message)
+class ArrowBatchTransformerTests(unittest.TestCase):
+    def test_flatten_struct_basic(self):
+        """Test flattening a struct column into separate columns."""
+        import pyarrow as pa
+
+        struct_array = pa.StructArray.from_arrays(
+            [pa.array([1, 2, 3]), pa.array(["a", "b", "c"])],
+            names=["x", "y"],
+        )
+        batch = pa.RecordBatch.from_arrays([struct_array], ["_0"])
+
+        flattened = ArrowBatchTransformer.flatten_struct(batch)
+
+        self.assertEqual(flattened.num_columns, 2)
+        self.assertEqual(flattened.column(0).to_pylist(), [1, 2, 3])
+        self.assertEqual(flattened.column(1).to_pylist(), ["a", "b", "c"])
+        self.assertEqual(flattened.schema.names, ["x", "y"])
+
+    def test_flatten_struct_empty_batch(self):
+        """Test flattening an empty batch."""
+        import pyarrow as pa
+
+        struct_type = pa.struct([("x", pa.int64()), ("y", pa.string())])
+        struct_array = pa.array([], type=struct_type)
+        batch = pa.RecordBatch.from_arrays([struct_array], ["_0"])
+
+        flattened = ArrowBatchTransformer.flatten_struct(batch)
+
+        self.assertEqual(flattened.num_rows, 0)
+        self.assertEqual(flattened.num_columns, 2)
+
+    def test_wrap_struct_basic(self):
+        """Test wrapping columns into a struct."""
+        import pyarrow as pa
+
+        batch = pa.RecordBatch.from_arrays(
+            [pa.array([1, 2, 3]), pa.array(["a", "b", "c"])],
+            names=["x", "y"],
+        )
+
+        wrapped = ArrowBatchTransformer.wrap_struct(batch)
+
+        self.assertEqual(wrapped.num_columns, 1)
+        self.assertEqual(wrapped.schema.names, ["_0"])
+
+        struct_col = wrapped.column(0)
+        self.assertEqual(len(struct_col), 3)
+        self.assertEqual(struct_col.field(0).to_pylist(), [1, 2, 3])
+        self.assertEqual(struct_col.field(1).to_pylist(), ["a", "b", "c"])
+
+    def test_wrap_struct_empty_columns(self):
+        """Test wrapping a batch with no columns."""
+        import pyarrow as pa
+
+        schema = pa.schema([])
+        batch = pa.RecordBatch.from_arrays([], schema=schema)
+
+        wrapped = ArrowBatchTransformer.wrap_struct(batch)
+
+        self.assertEqual(wrapped.num_columns, 1)
+        self.assertEqual(wrapped.num_rows, 0)
+
+    def test_wrap_struct_empty_batch(self):
+        """Test wrapping an empty batch with schema."""
+        import pyarrow as pa
+
+        schema = pa.schema([("x", pa.int64()), ("y", pa.string())])
+        batch = pa.RecordBatch.from_arrays(
+            [pa.array([], type=pa.int64()), pa.array([], type=pa.string())],
+            schema=schema,
+        )
+
+        wrapped = ArrowBatchTransformer.wrap_struct(batch)
+
+        self.assertEqual(wrapped.num_rows, 0)
+        self.assertEqual(wrapped.num_columns, 1)
 
 
 @unittest.skipIf(not have_pyarrow, pyarrow_requirement_message)
